less hostile language on the notification text, moved functions around
